+++ handlers/aws/handler.py
@@ -22,13 +22,13 @@ from .s3_sqs_trigger import _handle_s3_sqs_continuation, _handle_s3_sqs_event
 from .sqs_trigger import _handle_sqs_continuation, _handle_sqs_event
 from .utils import (
     CONFIG_FROM_PAYLOAD,
+    INTEGRATION_SCOPE_GENERIC,
     ConfigFileException,
     TriggerTypeException,
     capture_serverless,
     config_yaml_from_payload,
     config_yaml_from_s3,
     delete_sqs_record,
-    discover_integration_scope,
     expand_event_list_from_field_resolver,
     get_continuing_original_input_type,
     get_log_group_arn_and_region_from_log_group_name,
@@ -70,7 +70,7 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
         raise ConfigFileException("Empty config")
 
     try:
-        config = parse_config(config_yaml, _expanders, discover_integration_scope)
+        config = parse_config(config_yaml, _expanders)
     except Exception as e:
         raise ConfigFileException(e)
 
@@ -85,29 +85,32 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
         for replay_record in lambda_event["Records"]:
             event = json_parser(replay_record["body"])
             input_id = event["event_input_id"]
-            if input_id not in shipper_cache:
+            output_type = event["output_type"]
+            shipper_id = input_id + output_type
+            if shipper_id not in shipper_cache:
                 shipper = get_shipper_for_replay_event(
                     config=config,
-                    output_type=event["output_type"],
+                    output_type=output_type,
                     output_args=event["output_args"],
-                    event_input_id=event["event_input_id"],
+                    event_input_id=input_id,
                     replay_handler=replay_handler,
                 )
 
                 if shipper is None:
-                    shared_logger.warn(
+                    shared_logger.warning(
                         "no shipper for output in replay queue",
                         extra={"output_type": event["output_type"], "event_input_id": event["event_input_id"]},
                     )
                     continue
 
-                shipper_cache[input_id] = shipper
+                shipper_cache[shipper_id] = shipper
             else:
-                shipper = shipper_cache[input_id]
+                shipper = shipper_cache[shipper_id]
 
             shipper.send(event["event_payload"])
-            replay_handler.add_event_id_with_receipt_handle(
-                event_id=event["event_payload"]["_id"], receipt_handle=replay_record["receiptHandle"]
+            event_uniq_id: str = event["event_payload"]["_id"] + output_type
+            replay_handler.add_event_with_receipt_handle(
+                event_uniq_id=event_uniq_id, receipt_handle=replay_record["receiptHandle"]
             )
 
             if lambda_context is not None and lambda_context.get_remaining_time_in_millis() < _completion_grace_period:
@@ -145,14 +148,10 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
 
                 return "completed"
 
-        composite_shipper = get_shipper_from_input(
-            event_input=event_input, lambda_event=lambda_event, at_record=0, config_yaml=config_yaml
-        )
+        composite_shipper = get_shipper_from_input(event_input=event_input, config_yaml=config_yaml)
 
-        field_to_expand_event_list_from = event_input.expand_event_list_from_field
-        integration_scope = event_input.discover_integration_scope(lambda_event=lambda_event, at_record=0)
         expand_event_list_from_field = ExpandEventListFromField(
-            field_to_expand_event_list_from, integration_scope, expand_event_list_from_field_resolver
+            event_input.expand_event_list_from_field, INTEGRATION_SCOPE_GENERIC, expand_event_list_from_field_resolver
         )
 
         for (
@@ -218,15 +217,10 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
 
             return "completed"
 
-        composite_shipper = get_shipper_from_input(
-            event_input=event_input, lambda_event=lambda_event, at_record=0, config_yaml=config_yaml
-        )
+        composite_shipper = get_shipper_from_input(event_input=event_input, config_yaml=config_yaml)
 
-        integration_scope = event_input.discover_integration_scope(lambda_event=lambda_event, at_record=0)
-
-        field_to_expand_event_list_from = event_input.expand_event_list_from_field
         expand_event_list_from_field = ExpandEventListFromField(
-            field_to_expand_event_list_from, integration_scope, expand_event_list_from_field_resolver
+            event_input.expand_event_list_from_field, INTEGRATION_SCOPE_GENERIC, expand_event_list_from_field_resolver
         )
 
         for (
@@ -391,19 +385,9 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
             if input_id in composite_shipper_cache:
                 composite_shipper = composite_shipper_cache[input_id]
             else:
-                composite_shipper = get_shipper_from_input(
-                    event_input=event_input,
-                    lambda_event=lambda_event,
-                    at_record=current_sqs_record,
-                    config_yaml=config_yaml,
-                )
+                composite_shipper = get_shipper_from_input(event_input=event_input, config_yaml=config_yaml)
 
                 composite_shipper_cache[event_input.id] = composite_shipper
-
-            field_to_expand_event_list_from = event_input.expand_event_list_from_field
-            integration_scope = event_input.discover_integration_scope(
-                lambda_event=lambda_event, at_record=current_sqs_record
-            )
 
             sqs_record_body: dict[str, Any] = {}
             continuing_event_expanded_offset: Optional[int] = None
@@ -415,23 +399,18 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
                     sqs_record["messageAttributes"]["originalLastEventExpandedOffset"]["stringValue"]
                 )
 
-            if event_input.type == "s3-sqs":
-                sqs_record_body = json_parser(sqs_record["body"])
-                if "last_event_expanded_offset" in sqs_record_body["Records"][0]:
-                    continuing_event_expanded_offset = sqs_record_body["Records"][0]["last_event_expanded_offset"]
-
-            expand_event_list_from_field = ExpandEventListFromField(
-                field_to_expand_event_list_from,
-                integration_scope,
-                expand_event_list_from_field_resolver,
-                continuing_event_expanded_offset,
-            )
-
             if (
                 event_input.type == "kinesis-data-stream"
                 or event_input.type == "sqs"
                 or event_input.type == "cloudwatch-logs"
             ):
+                expand_event_list_from_field = ExpandEventListFromField(
+                    event_input.expand_event_list_from_field,
+                    INTEGRATION_SCOPE_GENERIC,
+                    expand_event_list_from_field_resolver,
+                    continuing_event_expanded_offset,
+                )
+
                 for es_event, last_ending_offset, last_event_expanded_offset in _handle_sqs_event(
                     sqs_record,
                     input_id,
@@ -468,10 +447,11 @@ def lambda_handler(lambda_event: dict[str, Any], lambda_context: context_.Contex
                         return "continuing"
 
             elif event_input.type == "s3-sqs":
+                sqs_record_body = json_parser(sqs_record["body"])
                 for es_event, last_ending_offset, last_event_expanded_offset, current_s3_record in _handle_s3_sqs_event(
                     sqs_record_body,
                     event_input.id,
-                    expand_event_list_from_field,
+                    event_input.expand_event_list_from_field,
                     event_input.json_content_type,
                     event_input.get_multiline_processor(),
                 ):