+++ share/__init__.py
@@ -2,7 +2,8 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 
-from .config import Config, ElasticsearchOutput, Input, Output, parse_config
+from .config import Config, ElasticsearchOutput, Input, LogstashOutput, Output, parse_config
+from .events import normalise_event
 from .expand_event_list_from_field import ExpandEventListFromField
 from .factory import MultilineFactory
 from .include_exlude import IncludeExcludeFilter, IncludeExcludeRule
@@ -10,3 +11,4 @@ from .json import json_dumper, json_parser
 from .logger import logger as shared_logger
 from .multiline import CollectBuffer, CountMultiline, PatternMultiline, ProtocolMultiline, WhileMultiline
 from .secretsmanager import aws_sm_expander
+from .utils import get_hex_prefix