+++ tests/share/test_config.py
@@ -3,7 +3,6 @@
 # you may not use this file except in compliance with the Elastic License 2.0.
 
 import re
-from typing import Any
 from unittest import TestCase
 
 import pytest
@@ -15,6 +14,7 @@ from share import (
     IncludeExcludeFilter,
     IncludeExcludeRule,
     Input,
+    LogstashOutput,
     Output,
     PatternMultiline,
     WhileMultiline,
@@ -31,7 +31,9 @@ class DummyOutput(Output):
 class TestOutput(TestCase):
     def test_init(self) -> None:
         with self.subTest("not valid type"):
-            with self.assertRaisesRegex(ValueError, "^`type` must be one of elasticsearch: another-type given$"):
+            with self.assertRaisesRegex(
+                ValueError, "^`type` must be one of elasticsearch,logstash: another-type given$"
+            ):
                 DummyOutput(output_type="another-type")
 
         with self.subTest("type not str"):
@@ -418,6 +420,75 @@ class TestElasticsearchOutput(TestCase):
 
 
 @pytest.mark.unit
+class TestLogstashOutput(TestCase):
+    def test_init(self) -> None:
+        with self.subTest("valid init with valid url"):
+            logstash = LogstashOutput(
+                logstash_url="http://localhost:8080",
+                tags=["tag1"],
+            )
+
+            assert logstash.type == "logstash"
+            assert logstash.logstash_url == "http://localhost:8080"
+        with self.subTest("valid init with valid url, max_batch_size and compression_level"):
+            logstash = LogstashOutput(
+                logstash_url="http://localhost:8080",
+                max_batch_size=400,
+                compression_level=2,
+                tags=["tag1"],
+            )
+
+            assert logstash.type == "logstash"
+            assert logstash.logstash_url == "http://localhost:8080"
+            assert logstash.max_batch_size == 400
+            assert logstash.compression_level == 2
+
+        with self.subTest("logstash_url not string"):
+            with self.assertRaisesRegex(ValueError, "`logstash_url` must be provided as string"):
+                LogstashOutput(logstash_url=0)  # type: ignore
+
+        with self.subTest("valid init with valid url, max_batch_size must be provided as int"):
+            with self.assertRaisesRegex(ValueError, "`max_batch_size` must be provided as int"):
+                LogstashOutput(
+                    logstash_url="http://localhost:8080",
+                    max_batch_size="string",  # type: ignore
+                )
+
+        with self.subTest("valid init with valid url, compression_level must be provided as int"):
+            with self.assertRaisesRegex(ValueError, "`compression_level` must be provided as int"):
+                LogstashOutput(
+                    logstash_url="http://localhost:8080",
+                    compression_level="string",  # type: ignore
+                )
+
+        with self.subTest("empty password"):
+            with self.assertRaisesRegex(ValueError, "`password` must be set when using `username`"):
+                LogstashOutput(logstash_url="http://localhost:8080", username="username")
+
+        with self.subTest("username not str"):
+            with self.assertRaisesRegex(ValueError, "`username` must be provided as string"):
+                LogstashOutput(
+                    logstash_url="http://localhost:8080",
+                    username=0,  # type:ignore
+                    password="password",
+                )
+
+        with self.subTest("password not str"):
+            with self.assertRaisesRegex(ValueError, "`password` must be provided as string"):
+                LogstashOutput(
+                    logstash_url="http://localhost:8080",
+                    username="username",
+                    password=0,  # type:ignore
+                )
+        with self.subTest("ssl_assert_fingerprint not str"):
+            with self.assertRaisesRegex(ValueError, "`ssl_assert_fingerprint` must be provided as string"):
+                LogstashOutput(
+                    logstash_url="http://localhost:8080",
+                    ssl_assert_fingerprint=0,  # type:ignore
+                )
+
+
+@pytest.mark.unit
 class TestInput(TestCase):
     def test_init(self) -> None:
         with self.subTest("valid s3-sqs init"):
@@ -526,6 +597,15 @@ class TestInput(TestCase):
 
             assert isinstance(input_sqs.get_output_by_type(output_type="elasticsearch"), ElasticsearchOutput)
 
+        with self.subTest("logstash output"):
+            input_sqs = Input(input_type="s3-sqs", input_id="id")
+            input_sqs.add_output(
+                output_type="logstash",
+                logstash_url="logstash_url",
+            )
+
+            assert isinstance(input_sqs.get_output_by_type(output_type="logstash"), LogstashOutput)
+
     def test_add_output(self) -> None:
         with self.subTest("elasticsearch output"):
             input_sqs = Input(input_type="s3-sqs", input_id="id")
@@ -541,9 +621,23 @@ class TestInput(TestCase):
 
             assert isinstance(input_sqs.get_output_by_type(output_type="elasticsearch"), ElasticsearchOutput)
 
-        with self.subTest("not elasticsearch output"):
+        with self.subTest("logstash output"):
             input_sqs = Input(input_type="s3-sqs", input_id="id")
-            with self.assertRaisesRegex(ValueError, "^`type` must be one of elasticsearch: another-type given$"):
+            input_sqs.add_output(
+                output_type="logstash",
+                logstash_url="logstash_url",
+                username="username",
+                password="password",
+                ssl_assert_fingerprint="fingerprint",
+            )
+
+            assert isinstance(input_sqs.get_output_by_type(output_type="logstash"), LogstashOutput)
+
+        with self.subTest("not elasticsearch or logstash output"):
+            input_sqs = Input(input_type="s3-sqs", input_id="id")
+            with self.assertRaisesRegex(
+                ValueError, "^`type` must be one of elasticsearch,logstash: another-type given$"
+            ):
                 input_sqs.add_output(output_type="another-type")
 
         with self.subTest("type is not str"):
@@ -613,22 +707,6 @@ class TestInput(TestCase):
             input_sqs = Input(input_type="s3-sqs", input_id="id")
             with self.assertRaisesRegex(KeyError, "'type"):
                 input_sqs.delete_output_by_type("type")
-
-    def test_discover_integration_scope(self) -> None:
-        with self.subTest("integration_scope_discoverer is None"):
-            input_sqs = Input(input_type="s3-sqs", input_id="id")
-            assert input_sqs.discover_integration_scope({}, 0) == ""
-
-        with self.subTest("integration_scope_discoverer is not none"):
-
-            def discover_integration_scope(lambda_event: dict[str, Any], at_record: int) -> str:
-                return "an_integration_scope"
-
-            input_sqs = Input(
-                input_type="s3-sqs", input_id="id", integration_scope_discoverer=discover_integration_scope
-            )
-
-            assert input_sqs.discover_integration_scope({}, 0) == "an_integration_scope"
 
 
 @pytest.mark.unit
@@ -825,7 +903,7 @@ class TestParseConfig(TestCase):
             with self.assertRaisesRegex(
                 ValueError,
                 "^An error occurred while applying output configuration at position 1 for input id: "
-                "`type` must be one of elasticsearch: another-type given$",
+                "`type` must be one of elasticsearch,logstash: another-type given$",
             ):
                 parse_config(
                     config_yaml="""